Fix error in AnyZerg.move() and freezing CombatZergs
There was a logic error in the AnyZerg move() method, causing the issue in
Movement.vacatePos(). The method vacatePos() sets creep.blockMovement to true,
but is meant to override it with force: true. The check in AnyZerg said
"!this.blockMovement && !force" therefore this method was not working, and the
creep not moving. Swapped it to "!this.blockMovement || force".

Also fixed an issue where Zergs of a role would stop working if one of their
kind was spawning, there was a return instead of an continue, aborting any Zerg
actions after it.

Squashed from eb2b78c9013411fd1b309efc98f286291f3a8b42